added readme file
A readme (or read me) file contains information about other files in a directory or archive and is commonly distributed with computer software, forming part of its documentation. Such a file is usually a text file called README.TXT, README.md, README.1ST, READ.ME, or simply README, although some Microsoft Windows software may occasionally include a README.WRI, README.RTF, or README.DOC. The name is chosen so that users unaware of the existence of this type of file would be drawn to read it.